moving ttvfaster to its own file
